refactor: change get booking url
change bookings/:id to /intern/:id/bookings

add get_booking url  to get the booking by id

refactor all bookings controller in one folder
