+++ client/src/constants/apiRoutes.js
@@ -30,14 +30,17 @@ const API_INTERN_COMPLETE_PROFILE = `${API_PREFIX}${INTERN_COMPLETE_PROFILE}`;
 const BOOKING_REQUEST_URL = "/new-booking";
 const API_BOOKING_REQUEST_URL = `${API_PREFIX}${BOOKING_REQUEST_URL}`;
 
+const GET_USER_BOOKINGS_URL = "/intern/:id/bookings";
+const API_GET_USER_BOOKINGS_URL = `${API_PREFIX}${GET_USER_BOOKINGS_URL}`;
+
+const GET_BOOKING_URL = "/bookings/:id";
+const API_GET_BOOKING_URL = `${API_PREFIX}${GET_BOOKING_URL}`;
+
 const SEARCH_PROFILES_URL = "/search-profiles";
 const API_SEARCH_PROFILES_URL = `${API_PREFIX}${SEARCH_PROFILES_URL}`;
 
 const ADMIN_STATS_URL = "/admin-stats";
 const API_ADMIN_STATS_URL = `${API_PREFIX}${ADMIN_STATS_URL}`;
-
-const GET_BOOKINGS_URL = "/bookings/:id";
-const API_GET_BOOKINGS_URL = `${API_PREFIX}${GET_BOOKINGS_URL}`;
 
 const VERIFY_PROFILE_URL = "/verify-profile";
 const API_VERIFY_PROFILE_URL = `${API_PREFIX}${VERIFY_PROFILE_URL}`;
@@ -73,9 +76,10 @@ module.exports = {
   API_HOST_PROFILE_URL,
   API_INTERN_PROFILE_URL,
   API_BOOKING_REQUEST_URL,
+  API_GET_USER_BOOKINGS_URL,
+  API_GET_BOOKING_URL,
   API_SEARCH_PROFILES_URL,
   API_ADMIN_STATS_URL,
-  API_GET_BOOKINGS_URL,
   API_VERIFY_PROFILE_URL,
   API_ORGS_DASHBOARD_URL,
   API_INTERN_DASHBOARD_URL,
@@ -95,9 +99,10 @@ module.exports = {
   HOST_PROFILE_URL,
   INTERN_PROFILE_URL,
   BOOKING_REQUEST_URL,
+  GET_USER_BOOKINGS_URL,
+  GET_BOOKING_URL,
   SEARCH_PROFILES_URL,
   ADMIN_STATS_URL,
-  GET_BOOKINGS_URL,
   VERIFY_PROFILE_URL,
   ORGS_DASHBOARD,
   INTERN_DASHBOARD_URL,
