+++ server/controllers/booking/index.js
@@ -1,0 +1,5 @@
+const newBookingRequest = require("./newBookingRequest");
+const getUserBookings = require("./getUserBookings");
+const viewBooking = require("./viewBooking");
+
+module.exports = { newBookingRequest, getUserBookings, viewBooking };
