+++ server/controllers/booking/viewBooking.js
@@ -1,0 +1,4 @@
+module.exports = (req, res, next) => {
+  console.log(req.params);
+  res.json({ data: req.body });
+};
